requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
passlib[bcrypt]
python-jose
psycopg2-binary
